Enhancements: advanced & bulk search, performance
This adds new features to search-all that makes the plugin
work better when there are lots of full-text enabled tables.

It also enables SQlite3's advanced search ("AND" and "OR" etc)
queries and adds a new bulk search interface that will run
queries from a CSV and output a CSV of all the search results.